+++ datafusion/physical-expr/src/equivalence/mod.rs
@@ -1,0 +1,9 @@
+mod class;
+mod ordering;
+mod projection;
+mod properties;
+
+pub use class::{EquivalenceClass, EquivalenceGroup};
+pub use ordering::OrderingEquivalenceClass;
+pub use projection::ProjectionMapping;
+pub use properties::EquivalenceProperties;
